[FIX] account_invoice_inter_company: Move demo data to test
This demo data is quite dangerous by design, since it creates compenies, accounts, decides the companies country, etc. All of that by hand.

In a tested integration environment, this fails when installing account-related addons due to companies not having a chart of accounts.

The fix is simple: remove this from demo data and leave it just as test data.