[FIX] base_multi_company: Proper _check_company when multiple companies selected
Steps to reproduce:

- Create two companies C1 and C2.
- Create a product with C1 and C2 as allowed companies.
- Having C1 as main company, update quantity on hand to 100.
- Switch to C2 as main company, but check the C1 company to be seen as
  well.
- Now update quantity on hand to any other value for the entry of the
  company C1.

Result:

'Stock>Inventory adjustment' belongs to company 'C2' and 'Product'
(product_id: 'Product 1') belongs to another company.

Explanation:

This is because the computed field `company_id` that is used for
checking the product company vs the stock.move company (the error is not
accurate, as it says the location, but it's the move), is returning the
main company where the user is.

Solution:

Inject a context whenever we run `_check_company` for having available
the company involved, and use that company as value for the record being
checked if included in the allowed companies.