[IMP-FIX] partner_multi_company: no longer define display_name as store
In the partner_multi_company module, the display_name field had to be
redefined as store and index because the module installation removed
those field properties. As of v17 the field is no longer store so it is
not necessary to redefine it as storable. In addition, this redefinition
causes errors when printing reports because the field is not calculated
when accessing and therefore only shows the data stored in the database,
which is not updated until one of the fields of the dependency changes.