[IMP] mail_template_multi_company: Enforce rules to XMLIDs
This is needed because mail templates are not expected to be inaccessible, so they are looked for with `self.env.ref` and used right away.
But if a template is in another company, it is found but an AccessError is raised as soon as any field is read.
With this change, it is not found in the first place.