[IMP] product_supplier_inter-multi_company: multi-company supplierinfo
Add `company_ids` to supplierinfo in order to make them visible to some
but not all companies. This solves an issue with supplierinfo for
products shared only among some companies.

When creating intercompany supplierinfo automatically,
sync the product's `company_ids` with the `supplierinfo`.

Steps to reproduce:

- product with companies A and B and not C
- generate an intercompany supplierinfo
- With only company C, try loading the supplierinfo list view ->
The supplierinfo is visible to company C, even though the product is not.